+++ tomcat/src/main/java/org/apache/coyote/http11/Controller.java
@@ -1,5 +1,8 @@
 package org.apache.coyote.http11;
 
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+
 public interface Controller {
 
     HttpResponse handle(HttpRequest request);
