+++ tomcat/src/main/java/org/apache/coyote/http11/controller/ErrorController.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11.controller;
 
 import org.apache.coyote.http11.Controller;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public class ErrorController implements Controller {
